+++ src/cli/__init__.py
@@ -56,6 +56,7 @@ from ._doctor_checks import (
     check_memory,
     check_mount_tmp,
     check_networking,
+    check_selinux,
     check_ssh,
     check_storage_driver,
     check_tmp_free_space,
@@ -68,6 +69,7 @@ from ._validators import (
     validate_dcos_configuration,
     validate_path_is_directory,
     validate_path_pair,
+    validate_volumes,
 )
 
 
@@ -297,6 +299,54 @@ def dcos_docker(verbose: None) -> None:
         'not set.'
     ),
 )
+@click.option(
+    '--custom-volume',
+    type=str,
+    callback=validate_volumes,
+    help=(
+        'Bind mount a volume on all cluster node containers. '
+        'See '
+        'https://docs.docker.com/engine/reference/run/#volume-shared-filesystems '  # noqa: E501
+        'for the syntax to use.'
+    ),
+    multiple=True,
+)
+@click.option(
+    '--custom-master-volume',
+    type=str,
+    callback=validate_volumes,
+    help=(
+        'Bind mount a volume on all cluster master node containers. '
+        'See '
+        'https://docs.docker.com/engine/reference/run/#volume-shared-filesystems '  # noqa: E501
+        'for the syntax to use.'
+    ),
+    multiple=True,
+)
+@click.option(
+    '--custom-agent-volume',
+    type=str,
+    callback=validate_volumes,
+    help=(
+        'Bind mount a volume on all cluster agent node containers. '
+        'See '
+        'https://docs.docker.com/engine/reference/run/#volume-shared-filesystems '  # noqa: E501
+        'for the syntax to use.'
+    ),
+    multiple=True,
+)
+@click.option(
+    '--custom-public-agent-volume',
+    type=str,
+    callback=validate_volumes,
+    help=(
+        'Bind mount a volume on all cluster public agent node containers. '
+        'See '
+        'https://docs.docker.com/engine/reference/run/#volume-shared-filesystems '  # noqa: E501
+        'for the syntax to use.'
+    ),
+    multiple=True,
+)
 def create(
     agents: int,
     artifact: str,
@@ -312,6 +362,10 @@ def create(
     copy_to_master: List[Tuple[Path, Path]],
     genconf_dir: Optional[Path],
     workspace_dir: Optional[Path],
+    custom_volume: Optional[Dict[str, Dict[str, str]]] = None,
+    custom_master_volume: Optional[Dict[str, Dict[str, str]]] = None,
+    custom_agent_volume: Optional[Dict[str, Dict[str, str]]] = None,
+    custom_public_agent_volume: Optional[Dict[str, Dict[str, str]]] = None,
 ) -> None:
     """
     Create a DC/OS cluster.
@@ -345,10 +399,6 @@ def create(
             \b
             If none of these are set, ``license_key_contents`` is not given.
     """  # noqa: E501
-    custom_master_mounts = {}  # type: Dict[str, Dict[str, str]]
-    custom_agent_mounts = {}  # type: Dict[str, Dict[str, str]]
-    custom_public_agent_mounts = {}  # type: Dict[str, Dict[str, str]]
-
     base_workspace_dir = workspace_dir or Path(gettempdir())
     workspace_dir = base_workspace_dir / uuid.uuid4().hex
 
@@ -402,9 +452,10 @@ def create(
             files_to_copy_to_installer.append((genconf_file, relative_path))
 
     cluster_backend = Docker(
-        custom_master_mounts=custom_master_mounts,
-        custom_agent_mounts=custom_agent_mounts,
-        custom_public_agent_mounts=custom_public_agent_mounts,
+        custom_container_mounts=custom_volume,
+        custom_master_mounts=custom_master_volume,
+        custom_agent_mounts=custom_agent_volume,
+        custom_public_agent_mounts=custom_public_agent_volume,
         linux_distribution=LINUX_DISTRIBUTIONS[linux_distribution],
         docker_version=DOCKER_VERSIONS[docker_version],
         storage_driver=DOCKER_STORAGE_DRIVERS.get(docker_storage_driver),
@@ -952,6 +1003,7 @@ def doctor() -> None:
         check_memory,
         check_mount_tmp,
         check_networking,
+        check_selinux,
         check_ssh,
         check_storage_driver,
         check_tmp_free_space,