+++ CHANGELOG.rst
@@ -6,6 +6,18 @@ Changelog
 Next
 ----
 
+- Logs from dependencies are no longer emitted.
+
+2018.04.19.0
+------------
+
+- The AWS backend now supports DC/OS 1.9.
+- The Docker backend now supports having custom mounts which apply to all nodes.
+- Add ``custom-volume`` parameter (and similar for each node type) to ``dcos-docker create``.
+
+2018.04.11.0
+------------
+
 - Add an AWS backend to the library.
 - Add ability to control which labels are added to particular node types on the ``Docker`` backend.
 - Add support for Ubuntu on the ``Docker`` backend.